+++ functions/Inputs.php
@@ -813,7 +813,7 @@ function MultipleCheckboxInput( $value, $name, $title, $options, $extra = '', $d
  *
  * @since 5.6 Support option groups (`<optgroup>`) by adding 'group' to $extra.
  * @since 6.0 Support multiple values.
- * @since 10.8.4 Fix save multiple SelectInput() when none selected, add hidden empty input
+ * @since 10.8.4 Fix save multiple SelectInput() when none selected, add hidden empty input (only if $allow_na)
  *
  * @example SelectInput( $value, 'values[' . $id . '][' . $name . ']', '', $options, 'N/A', $extra )
  *
@@ -963,8 +963,10 @@ function SelectInput( $values, $name, $title = '', $options = [], $allow_na = 'N
 
 	$select .= '</select>' . FormatInputTitle( $title, $id, $required );
 
-	if ( $is_multiple )
+	if ( $is_multiple
+		&& $allow_na !== false )
 	{
+		// Fix save multiple SelectInput() when none selected, add hidden empty input (only if $allow_na)
 		$select .= '<input type="hidden" name="' . AttrEscape( $name ) . '" value="">';
 	}